flask/requirements.txt
langchain
unstructured
openai
pinecone-client
firebase-admin
flask
flask-HTTPAuth
